+++ src/components/form/choice/ChoiceItem.styled.ts
@@ -1,0 +1,9 @@
+import { styled } from "../../../theme";
+
+const StyledChoiceLabel = styled("label", {
+    "--StyledChoiceLabelDummyVar": "0",
+});
+
+export {
+    StyledChoiceLabel,
+};